+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Python v1.0.1 - 2025-04-29
+- Updated documentation.
+
+## Python v1.0.0 - 2025-04-29
+- Initial release of the Aristech TTS-Client for Python.
+
 ## NodeJS v1.1.0 - 2025-04-29
 - Updated dependencies.
 - Deprecated `process` in favor of `runFunctions` for better cohesion with the protobuf API.
@@ -8,8 +14,6 @@
 - Updated dependencies.
 - Removed unused dependencies.
 
-## Python v1.0.0 - 2024-10-29
-- Initial release of the Aristech TTS-Client for Python.
 
 ## Rust v1.0.0 - 2024-10-29
 - Initial release of the Aristech TTS-Client for Rust.